apps: split serialization outside apps util
