+++ pkg/apps/util/appserialization/scheme.go
@@ -1,4 +1,4 @@
-package util
+package appserialization
 
 import (
 	"k8s.io/apimachinery/pkg/runtime"
@@ -6,8 +6,6 @@ import (
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 
 	appsv1 "github.com/openshift/api/apps/v1"
-	"github.com/openshift/origin/pkg/api/legacy"
-	appsv1helpers "github.com/openshift/origin/pkg/apps/apis/apps/v1"
 )
 
 var (
@@ -21,12 +19,12 @@ var (
 )
 
 func init() {
-	legacy.InstallInternalLegacyApps(annotationDecodingScheme)
-	utilruntime.Must(appsv1helpers.Install(annotationDecodingScheme))
+	utilruntime.Must(appsv1.Install(annotationDecodingScheme))
+	utilruntime.Must(appsv1.DeprecatedInstallWithoutGroup(annotationDecodingScheme))
 	annotationDecoderCodecFactory := serializer.NewCodecFactory(annotationDecodingScheme)
 	annotationDecoder = annotationDecoderCodecFactory.UniversalDecoder(appsv1.GroupVersion)
 
-	utilruntime.Must(appsv1helpers.Install(annotationEncodingScheme))
+	utilruntime.Must(appsv1.Install(annotationEncodingScheme))
 	annotationEncoderCodecFactory := serializer.NewCodecFactory(annotationEncodingScheme)
 	annotationEncoder = annotationEncoderCodecFactory.LegacyCodec(appsv1.GroupVersion)
 }
