+++ pkg/apps/util/test/ok.go
@@ -1,15 +1,10 @@
 package test
 
 import (
-	"testing"
-
+	appsv1 "github.com/openshift/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/util/sets"
-	"k8s.io/kubernetes/pkg/api/legacyscheme"
-
-	appsv1 "github.com/openshift/api/apps/v1"
 )
 
 const (
@@ -47,24 +42,6 @@ func OkDeploymentConfigStatus(version int64) appsv1.DeploymentConfigStatus {
 	return appsv1.DeploymentConfigStatus{
 		LatestVersion: version,
 	}
-}
-
-func OkImageChangeDetails() *appsv1.DeploymentDetails {
-	return &appsv1.DeploymentDetails{
-		Causes: []appsv1.DeploymentCause{{
-			Type: appsv1.DeploymentTriggerOnImageChange,
-			ImageTrigger: &appsv1.DeploymentCauseImageTrigger{
-				From: corev1.ObjectReference{
-					Name: ImageStreamName + ":latest",
-					Kind: "ImageStreamTag",
-				}}}}}
-}
-
-func OkConfigChangeDetails() *appsv1.DeploymentDetails {
-	return &appsv1.DeploymentDetails{
-		Causes: []appsv1.DeploymentCause{{
-			Type: appsv1.DeploymentTriggerOnConfigChange,
-		}}}
 }
 
 func OkStrategy() appsv1.DeploymentStrategy {
@@ -173,24 +150,6 @@ func OkPodTemplate() *corev1.PodTemplateSpec {
 	}
 }
 
-func OkPodTemplateChanged() *corev1.PodTemplateSpec {
-	template := OkPodTemplate()
-	template.Spec.Containers[0].Image = DockerImageReference
-	return template
-}
-
-func OkPodTemplateMissingImage(missing ...string) *corev1.PodTemplateSpec {
-	set := sets.NewString(missing...)
-	template := OkPodTemplate()
-	for i, c := range template.Spec.Containers {
-		if set.Has(c.Name) {
-			// remember that slices use copies, so have to ref array entry explicitly
-			template.Spec.Containers[i].Image = ""
-		}
-	}
-	return template
-}
-
 func OkConfigChangeTrigger() appsv1.DeploymentTriggerPolicy {
 	return appsv1.DeploymentTriggerPolicy{
 		Type: appsv1.DeploymentTriggerOnConfigChange,
@@ -213,56 +172,7 @@ func OkImageChangeTrigger() appsv1.DeploymentTriggerPolicy {
 	}
 }
 
-func OkTriggeredImageChange() appsv1.DeploymentTriggerPolicy {
-	ict := OkImageChangeTrigger()
-	ict.ImageChangeParams.LastTriggeredImage = DockerImageReference
-	return ict
-}
-
-func OkNonAutomaticICT() appsv1.DeploymentTriggerPolicy {
-	ict := OkImageChangeTrigger()
-	ict.ImageChangeParams.Automatic = false
-	return ict
-}
-
-func OkTriggeredNonAutomatic() appsv1.DeploymentTriggerPolicy {
-	ict := OkNonAutomaticICT()
-	ict.ImageChangeParams.LastTriggeredImage = DockerImageReference
-	return ict
-}
-
 func TestDeploymentConfig(config *appsv1.DeploymentConfig) *appsv1.DeploymentConfig {
 	config.Spec.Test = true
 	return config
-}
-
-func RemoveTriggerTypes(config *appsv1.DeploymentConfig, triggerTypes ...appsv1.DeploymentTriggerType) {
-	types := sets.NewString()
-	for _, triggerType := range triggerTypes {
-		types.Insert(string(triggerType))
-	}
-
-	remaining := []appsv1.DeploymentTriggerPolicy{}
-	for _, trigger := range config.Spec.Triggers {
-		if types.Has(string(trigger.Type)) {
-			continue
-		}
-		remaining = append(remaining, trigger)
-	}
-
-	config.Spec.Triggers = remaining
-}
-
-func RoundTripConfig(t *testing.T, config *appsv1.DeploymentConfig) *appsv1.DeploymentConfig {
-	versioned, err := legacyscheme.Scheme.ConvertToVersion(config, appsv1.SchemeGroupVersion)
-	if err != nil {
-		t.Errorf("unexpected conversion error: %v", err)
-		return nil
-	}
-	defaulted, err := legacyscheme.Scheme.ConvertToVersion(versioned, appsv1.SchemeGroupVersion)
-	if err != nil {
-		t.Errorf("unexpected conversion error: %v", err)
-		return nil
-	}
-	return defaulted.(*appsv1.DeploymentConfig)
 }
